Split work lane smoke fixtures
